# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Learning": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_username": {
            "title": "Created By Username",
            "type": "string"
          },
          "embedded": {
            "title": "Embedded",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "path_glob": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Glob"
          },
          "pr_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Number"
          },
          "repo_full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Full Name"
          },
          "repo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Id"
          },
          "scope": {
            "enum": [
              "repo",
              "org"
            ],
            "title": "Scope",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "scope",
          "repo_id",
          "repo_full_name",
          "path_glob",
          "source_url",
          "pr_number",
          "created_by_username",
          "embedded",
          "created_at",
          "updated_at"
        ],
        "title": "Learning",
        "type": "object"
      },
      "LearningCreate": {
        "properties": {
          "path_glob": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Glob"
          },
          "repo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Id"
          },
          "scope": {
            "default": "repo",
            "enum": [
              "repo",
              "org"
            ],
            "title": "Scope",
            "type": "string"
          },
          "text": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "LearningCreate",
        "type": "object"
      },
      "LearningList": {
        "properties": {
          "learnings": {
            "items": {
              "$ref": "#/components/schemas/Learning"
            },
            "title": "Learnings",
            "type": "array"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          }
        },
        "required": [
          "learnings",
          "next_before"
        ],
        "title": "LearningList",
        "type": "object"
      },
      "LearningUpdate": {
        "properties": {
          "path_glob": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Glob"
          },
          "scope": {
            "anyOf": [
              {
                "enum": [
                  "repo",
                  "org"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "title": "LearningUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "HootPR API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/orgs/{org_slug}/learnings": {
      "get": {
        "operationId": "list_learnings_api_orgs__org_slug__learnings_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "repo_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repo Id"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Learnings"
      },
      "post": {
        "operationId": "create_learning_api_orgs__org_slug__learnings_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Learning"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Learning"
      }
    },
    "/api/orgs/{org_slug}/learnings/{learning_id}": {
      "delete": {
        "operationId": "delete_learning_api_orgs__org_slug__learnings__learning_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "learning_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learning Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Learning"
      },
      "patch": {
        "operationId": "update_learning_api_orgs__org_slug__learnings__learning_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "learning_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Learning Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Learning"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Learning"
      }
    },
...
  }
}